moved filter_utils of CMS_1JET_8TEV into legacy jet utils
