+++ nnpdf_data/nnpdf_data/new_commondata/CMS_2JET_7TEV/kinematics.yaml
@@ -3,7 +3,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 197.0
     mid: 246.5
     max: 296.0
@@ -15,7 +15,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 296.0
     mid: 357.5
     max: 419.0
@@ -27,7 +27,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 419.0
     mid: 492.0
     max: 565.0
@@ -39,7 +39,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 565.0
     mid: 652.5
     max: 740.0
@@ -51,7 +51,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 740.0
     mid: 842.0
     max: 944.0
@@ -63,7 +63,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 944.0
     mid: 1062.5
     max: 1181.0
@@ -75,7 +75,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 1181.0
     mid: 1318.0
     max: 1455.0
@@ -87,7 +87,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 1455.0
     mid: 1612.5
     max: 1770.0
@@ -99,7 +99,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 1770.0
     mid: 1951.0
     max: 2132.0
@@ -111,7 +111,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 2132.0
     mid: 2339.0
     max: 2546.0
@@ -123,7 +123,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 2546.0
     mid: 2782.5
     max: 3019.0
@@ -135,7 +135,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 3019.0
     mid: 3217.5
     max: 3416.0
@@ -147,7 +147,7 @@ bins:
     min: 0.0
     mid: 0.25
     max: 0.5
-  m12:
+  m_jj:
     min: 3416.0
     mid: 3713.0
     max: 4010.0
@@ -159,7 +159,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 270.0
     mid: 328.0
     max: 386.0
@@ -171,7 +171,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 386.0
     mid: 456.0
     max: 526.0
@@ -183,7 +183,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 526.0
     mid: 609.5
     max: 693.0
@@ -195,7 +195,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 693.0
     mid: 791.5
     max: 890.0
@@ -207,7 +207,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 890.0
     mid: 1004.0
     max: 1118.0
@@ -219,7 +219,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 1118.0
     mid: 1250.5
     max: 1383.0
@@ -231,7 +231,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 1383.0
     mid: 1535.0
     max: 1687.0
@@ -243,7 +243,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 1687.0
     mid: 1862.0
     max: 2037.0
@@ -255,7 +255,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 2037.0
     mid: 2237.5
     max: 2438.0
@@ -267,7 +267,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 2438.0
     mid: 2666.5
     max: 2895.0
@@ -279,7 +279,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 2895.0
     mid: 3155.5
     max: 3416.0
@@ -291,7 +291,7 @@ bins:
     min: 0.5
     mid: 0.75
     max: 1.0
-  m12:
+  m_jj:
     min: 3416.0
     mid: 3713.0
     max: 4010.0
@@ -303,7 +303,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 419.0
     mid: 492.0
     max: 565.0
@@ -315,7 +315,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 565.0
     mid: 652.5
     max: 740.0
@@ -327,7 +327,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 740.0
     mid: 842.0
     max: 944.0
@@ -339,7 +339,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 944.0
     mid: 1062.5
     max: 1181.0
@@ -351,7 +351,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 1181.0
     mid: 1318.0
     max: 1455.0
@@ -363,7 +363,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 1455.0
     mid: 1612.5
     max: 1770.0
@@ -375,7 +375,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 1770.0
     mid: 1951.0
     max: 2132.0
@@ -387,7 +387,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 2132.0
     mid: 2339.0
     max: 2546.0
@@ -399,7 +399,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 2546.0
     mid: 2846.5
     max: 3147.0
@@ -411,7 +411,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 3147.0
     mid: 3500.5
     max: 3854.0
@@ -423,7 +423,7 @@ bins:
     min: 1.0
     mid: 1.25
     max: 1.5
-  m12:
+  m_jj:
     min: 3854.0
     mid: 4181.5
     max: 4509.0
@@ -435,7 +435,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 565.0
     mid: 652.5
     max: 740.0
@@ -447,7 +447,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 740.0
     mid: 842.0
     max: 944.0
@@ -459,7 +459,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 944.0
     mid: 1062.5
     max: 1181.0
@@ -471,7 +471,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 1181.0
     mid: 1318.0
     max: 1455.0
@@ -483,7 +483,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 1455.0
     mid: 1612.5
     max: 1770.0
@@ -495,7 +495,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 1770.0
     mid: 1951.0
     max: 2132.0
@@ -507,7 +507,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 2132.0
     mid: 2339.0
     max: 2546.0
@@ -519,7 +519,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 2546.0
     mid: 2782.5
     max: 3019.0
@@ -531,7 +531,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 3019.0
     mid: 3288.5
     max: 3558.0
@@ -543,7 +543,7 @@ bins:
     min: 1.5
     mid: 1.75
     max: 2.0
-  m12:
+  m_jj:
     min: 3558.0
     mid: 4308.0
     max: 5058.0
@@ -555,7 +555,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 1000.0
     mid: 1123.0
     max: 1246.0
@@ -567,7 +567,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 1246.0
     mid: 1388.0
     max: 1530.0
@@ -579,7 +579,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 1530.0
     mid: 1693.0
     max: 1856.0
@@ -591,7 +591,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 1856.0
     mid: 2043.5
     max: 2231.0
@@ -603,7 +603,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 2231.0
     mid: 2445.0
     max: 2659.0
@@ -615,7 +615,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 2659.0
     mid: 2903.0
     max: 3147.0
@@ -627,7 +627,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 3147.0
     mid: 3425.5
     max: 3704.0
@@ -639,7 +639,7 @@ bins:
     min: 2.0
     mid: 2.25
     max: 2.5
-  m12:
+  m_jj:
     min: 3704.0
     mid: 4381.0
     max: 5058.0
